\name{importText1}
\alias{importText1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
%%  ~~importText1. ~~
}
\description{
%%  ~~ This is a function that allows R to take a text file from the directory and read each line ~~
}
\usage{
}
%- readLines(file.choose(x))
\arguments{
  \item{x}{
%%     ~~Select the file from the directory.~~
}
}

\value{
%%  ~imports txt file.
}

\author{
%%  ~~Stephy Oge~~
}

\examples{
 readLines(file.choose())
{
  }
}
