\name{inspectText}
\alias{inspectText}

\title{
%%  ~~Inspect the text file. ~~
}

\usage{
inspect(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
%%     ~~text file~~
}
}
\details{
%%  ~~ inspect() will inspect the corpus text ~~
}

\examples{
inspect(speech1)
{
  }
}
