\name{loadAsCorpus}
\alias{loadAsCorpus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
%%  ~~ Load as Corpus~~
}
\description{
%%  ~~ function loads information as corpus. ~~
}
\usage{
Corpus(VectorSource(x))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
%%     ~~a text file or .txt~~
}
}
\details{
%%  ~~ VectorSource() function creates a corpus of character vectors ~~
}
\value{
%%  ~nothing.
}

\examples{
 speech1 <- Corpus(VectorSource(speech))
{
  }
}
