\name{matrix}
\alias{matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
%%  ~~Term-document matrix. ~~
}
\description{
%%  ~~ place corpus text into a matrix. ~~
}
\usage{
TermDocumentMatrix(x)
}

\arguments{
  \item{x}{
%%     ~~text file~~
}
}

\examples{
 dtm <- TermDocumentMatrix(speech1)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
