\name{textTransformation4}
\alias{textTransformation4}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
%%  ~~ Text Transformation~~
}
\description{
%%  ~~ transform or replace unwanted characters with spaces. ~~
}
\usage{
content_transformer(x)
}
%- tm_map(x, toSpace,...) 
\arguments{
  \item{x}{
%%     ~~The corpus text file~~
}
}
\details{
%%  ~~ remove any special characters you do not want to appear in the word cloud ~~
}
\value{
%%  ~nothing.
}

\examples{
tm_map(speech1, toSpace, "@")
{
  }
}
