\name{wordCloud7}
\alias{wordCloud7}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
%%  ~~ wordCloud. ~~
}
\description{
%%  ~~ A white and black word cloud from a MLK speech to celebrate Black History Month. ~~
}
\usage{
set.seed(x)
}
%- wordcloud2::wordcloud(x,...)
\arguments{
  \item{x}{
%%     ~~insert the matrix or data frame containing words~~
}
}

\examples{
set.seed(1234)
wordcloud2::wordcloud2(D, size = 0.5, color = "white", backgroundColor = "black")
{
  }
}
